% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-attr_set.R
\name{add_attr.attr_set}
\alias{add_attr.attr_set}
\title{S3 Method - Add New Attributes to Attribute Set}
\usage{
\method{add_attr}{attr_set}(obj, x)
}
\arguments{
\item{obj}{S3 Object}

\item{x}{character}
}
\value{
character
}
\description{
S3 Method - Add New Attributes to Attribute Set
}
\examples{
\dontrun{
add_attr(attr_set, x)
}
}
