% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-fdep.R
\name{add_lhs.fdep}
\alias{add_lhs.fdep}
\title{S3 Method - Add New Attributes to LHS Attribute Set}
\usage{
\method{add_lhs}{fdep}(obj, x)
}
\arguments{
\item{obj}{S3 Object}

\item{x}{character}
}
\value{
character
}
\description{
S3 Method - Add New Attributes to LHS Attribute Set
}
\examples{
\dontrun{
add_lhs(fdep, x)
}
}
