% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-attr_set.R
\name{attr_set}
\alias{attr_set}
\title{S3 Helper Function for Class 'attr_set'}
\usage{
attr_set(attributes)
}
\arguments{
\item{attributes}{character}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'attr_set'
}
\examples{
\dontrun{
test <- attr_set(x)
}
}
