% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-err_func.R
\name{err_fd_from_records}
\alias{err_fd_from_records}
\title{Calculate 'Functional Dependency Error' from Position List Records}
\usage{
err_fd_from_records(pos, x_col, y_col)
}
\arguments{
\item{pos}{S3 Object of class 'pos_list'}

\item{x_col}{character - Vector of column names}

\item{y_col}{character - Vector of column names}
}
\value{
R Object
}
\description{
Calculate 'Functional Dependency Error' from Position List Records
}
\examples{
\dontrun{
result <- err_key_from_records(strip_part)
}
}
