% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-err_func.R
\name{err_fd_from_strip_part}
\alias{err_fd_from_strip_part}
\title{Calculate 'Functional Dependency Error' from Stripped Product}
\usage{
err_fd_from_strip_part(x_part, xy_part)
}
\arguments{
\item{x_part}{S3 Object}

\item{xy_part}{S3 Object}
}
\value{
R Object
}
\description{
Calculate 'Functional Dependency Error' from Stripped Product
}
\examples{
\dontrun{
result <- err_fd_from_strip_part(x_part, xy_part)
}
}
