% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-err_func.R
\name{err_key_from_records}
\alias{err_key_from_records}
\title{Calculate 'Key Error' from Position List Records}
\usage{
err_key_from_records(pos, col, err_only)
}
\arguments{
\item{pos}{S3 Object of class 'pos_list'}

\item{col}{character - Vector of column names}

\item{err_only}{TRUE/FALSE - Optionally specify if only failing values should be returned}
}
\value{
R Object
}
\description{
Calculate 'Key Error' from Position List Records
}
\examples{
\dontrun{
result <- err_key_from_records(pos_list, col_list)
}
}
