% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-err_func.R
\name{err_key_from_strip_part}
\alias{err_key_from_strip_part}
\title{Calculate 'Key Error' from Stripped Product}
\usage{
err_key_from_strip_part(part)
}
\arguments{
\item{part}{S3 Object}
}
\value{
R Object
}
\description{
Calculate 'Key Error' from Stripped Product
}
\examples{
\dontrun{
result <- err_key_from_strip_part(strip_part)
}
}
