% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-fdep.R
\name{fdep}
\alias{fdep}
\title{S3 Helper Function for Class 'fdep'}
\usage{
fdep(lhs, rhs)
}
\arguments{
\item{lhs}{S3 Object}

\item{rhs}{S3 Object}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'fdep'
}
\examples{
\dontrun{
result <- fdep(lhs, rhs)
}
}
