% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-frame_to_pl.R
\name{frame_to_pl}
\alias{frame_to_pl}
\title{Convert R DataFrame into a 'Position List'}
\usage{
frame_to_pl(data, index, records, ...)
}
\arguments{
\item{data}{data.frame}

\item{index}{TRUE/FALSE}

\item{records}{TRUE/FALSE}

\item{...}{ellipsis}
}
\value{
R Object
}
\description{
Convert R DataFrame into a 'Position List'
}
\examples{
\dontrun{
pos_list <- frame_to_pl(raw_data_frame, TRUE, TRUE)
}
}
