% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-func_dep.R
\name{func_dep}
\alias{func_dep}
\title{S3 Helper Function for Class 'func_dep'}
\usage{
func_dep(lhs, rhs)
}
\arguments{
\item{lhs}{S3 Object}

\item{rhs}{S3 Object}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'func_dep'
}
\examples{
\dontrun{
result <- func_dep(lhs, rhs)
}
}
