% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-frame_to_pl.R
\name{gen_pl_records}
\alias{gen_pl_records}
\title{Generate 'Records' component of a 'Position List'}
\usage{
gen_pl_records(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{ellipsis}
}
\value{
R Object
}
\description{
Generate 'Records' component of a 'Position List'
}
\examples{
\dontrun{
map <- gen_pl_records(raw_data_frame)
}
}
