% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-attr_set.R
\name{get_attr.attr_set}
\alias{get_attr.attr_set}
\title{S3 Method - Return Character Vector of Attributes}
\usage{
\method{get_attr}{attr_set}(obj)
}
\arguments{
\item{obj}{S3 Object}
}
\value{
character
}
\description{
S3 Method - Return Character Vector of Attributes
}
\examples{
\dontrun{
result <- get_attr(attr_set)
}
}
