% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-attr_set.R
\name{get_attr}
\alias{get_attr}
\title{S3 Generic - Return Character Vector of Attributes}
\usage{
get_attr(obj)
}
\arguments{
\item{obj}{S3 Object}
}
\value{
character
}
\description{
S3 Generic - Return Character Vector of Attributes
}
\examples{
\dontrun{
result <- get_attr(attr_set)
}
}
