% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-fdep.R
\name{get_lhs.fdep}
\alias{get_lhs.fdep}
\title{S3 Method - Return Character Vector of LHS Attributes}
\usage{
\method{get_lhs}{fdep}(obj)
}
\arguments{
\item{obj}{S3 Object}
}
\value{
character
}
\description{
S3 Method - Return Character Vector of LHS Attributes
}
\examples{
\dontrun{
result <- get_lhs(fdep)
}
}
