% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-fdep.R
\name{get_lhs}
\alias{get_lhs}
\title{S3 Generic - Return Character Vector of LHS Attributes}
\usage{
get_lhs(obj)
}
\arguments{
\item{obj}{S3 Object}
}
\value{
character
}
\description{
S3 Generic - Return Character Vector of LHS Attributes
}
\examples{
\dontrun{
result <- get_lhs(fdep)
}
}
