% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-fdep.R
\name{get_rhs.fdep}
\alias{get_rhs.fdep}
\title{S3 Method - Return Character Vector of RHS Attributes}
\usage{
\method{get_rhs}{fdep}(obj)
}
\arguments{
\item{obj}{S3 Object}
}
\value{
character
}
\description{
S3 Method - Return Character Vector of RHS Attributes
}
\examples{
\dontrun{
result <- get_rhs(fdep)
}
}
