% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-fdep.R
\name{get_rhs}
\alias{get_rhs}
\title{S3 Generic - Return Character Vector of RHS Attributes}
\usage{
get_rhs(obj)
}
\arguments{
\item{obj}{S3 Object}
}
\value{
character
}
\description{
S3 Generic - Return Character Vector of RHS Attributes
}
\examples{
\dontrun{
result <- get_rhs(fdep)
}
}
