% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-pos_list.R
\name{key_err}
\alias{key_err}
\title{S3 Generic 'key_err'}
\usage{
key_err(obj, col, err_only)
}
\arguments{
\item{obj}{S3 Object}

\item{col}{character - Vector of column names used to create partition}

\item{err_only}{TRUE/FALSE - Optionally specify if only failing values should be returned}
}
\value{
list
}
\description{
S3 Generic 'key_err'
}
\examples{
\dontrun{
test <- strip_part(pos_list)
}
}
