% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-attr_set.R
\name{new_attr_set}
\alias{new_attr_set}
\title{S3 Constructor for Class 'attr_set'}
\usage{
new_attr_set(attr)
}
\arguments{
\item{attr}{character}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'attr_set'
}
\examples{
\dontrun{
result <- new_attr_set(x)
}
}
