% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-fdep.R
\name{new_fdep}
\alias{new_fdep}
\title{S3 Constructor for Class 'fdep'}
\usage{
new_fdep(lhs, rhs)
}
\arguments{
\item{lhs}{S3 Object - attr_set}

\item{rhs}{S3 Object - attr_set}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'fdep'
}
\examples{
\dontrun{
test <- new_fdep(lhs, rhs)
}
}
