% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-func_dep.R
\name{new_func_dep}
\alias{new_func_dep}
\title{S3 Constructor for Class 'func_dep'}
\usage{
new_func_dep(lhs, rhs)
}
\arguments{
\item{lhs}{S3 Object - attr_set}

\item{rhs}{S3 Object - attr_set}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'func_dep'
}
\examples{
\dontrun{
test <- new_func_dep(lhs, rhs)
}
}
