% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-pos_list.R
\name{new_pos_list}
\alias{new_pos_list}
\title{S3 Constructor for Class 'pos_list'}
\usage{
new_pos_list(frame)
}
\arguments{
\item{frame}{data.frame}
}
\value{
S3 Object
}
\description{
S3 Constructor for Class 'pos_list'
}
\examples{
\dontrun{
test <- new_pos_list(raw_data_frame)
}
}
