% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-frame_to_pl.R
\name{pl_to_frame}
\alias{pl_to_frame}
\title{Convert a 'Position List' into R DataFrame}
\usage{
pl_to_frame(pl, use_index)
}
\arguments{
\item{pl}{R Object}

\item{use_index}{TRUE/FALSE}
}
\value{
data.frame
}
\description{
Convert a 'Position List' into R DataFrame
}
\examples{
\dontrun{
new_df <- pl_to_frame(pl, TRUE)
}
}
