% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-pos_list.R
\name{pos_list}
\alias{pos_list}
\title{S3 Helper Function for Class 'pos_list'}
\usage{
pos_list(frame)
}
\arguments{
\item{frame}{data.frame}
}
\value{
S3 Object
}
\description{
S3 Helper Function for Class 'pos_list'
}
\examples{
\dontrun{
test <- pos_list(IMF_DATA)
}
}
