% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-attr_set.R
\name{rm_attr.attr_set}
\alias{rm_attr.attr_set}
\title{S3 Method - Remove Attributes from Attribute Set}
\usage{
\method{rm_attr}{attr_set}(obj, x)
}
\arguments{
\item{obj}{S3 Object}

\item{x}{character}
}
\value{
character
}
\description{
S3 Method - Remove Attributes from Attribute Set
}
\examples{
\dontrun{
rm_attr(attr_set, x)
}
}
