% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-attr_set.R
\name{rm_attr}
\alias{rm_attr}
\title{S3 Generic - Remove Attributes from Attribute Set}
\usage{
rm_attr(obj, x)
}
\arguments{
\item{obj}{S3 Object}

\item{x}{character}
}
\value{
character
}
\description{
S3 Generic - Remove Attributes from Attribute Set
}
\examples{
\dontrun{
rm_attr(attr_set, x)
}
}
