% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-pos_list.R
\name{strip_part.pos_list}
\alias{strip_part.pos_list}
\title{S3 Generic 'strip_part' returns a stripped partition}
\usage{
\method{strip_part}{pos_list}(obj, ...)
}
\arguments{
\item{obj}{S3 Object}

\item{...}{ellipsis}
}
\value{
list
}
\description{
S3 Generic 'strip_part' returns a stripped partition
}
\examples{
\dontrun{
test <- strip_part.pos_list(pos_list)
}
}
