% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods-pos_list.R
\name{strip_part}
\alias{strip_part}
\title{S3 Generic 'strip_part'}
\usage{
strip_part(obj, col)
}
\arguments{
\item{obj}{S3 Object}

\item{col}{character - Vector of column names used to create partition}
}
\value{
list
}
\description{
S3 Generic 'strip_part'
}
\examples{
\dontrun{
test <- strip_part(pos_list)
}
}
