% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-strip_part.R
\name{strip_part_equal}
\alias{strip_part_equal}
\title{Test Stripped Partitions for Equality}
\usage{
strip_part_equal(x, y)
}
\arguments{
\item{x}{Stripped Partition}

\item{y}{Stripped Partition}
}
\value{
R Object
}
\description{
Test Stripped Partitions for Equality
}
\examples{
\dontrun{
test <- strip_part_equal(x,y)
}
}
