% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-strip_part.R
\name{strip_part_from_record}
\alias{strip_part_from_record}
\title{Generate 'Stripped Partition' from Position List Records}
\usage{
strip_part_from_record(pos, col)
}
\arguments{
\item{pos}{S3 Object of class 'pos_list'}

\item{col}{character - Vector of column names used to create partition}
}
\value{
S3 Object
}
\description{
Generate 'Stripped Partition' from Position List Records
}
\examples{
\dontrun{
output <- strip_part_from_record(pos_list, column_names)
}
}
