% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-strip_part.R
\name{strip_product_r}
\alias{strip_product_r}
\title{Generate 'Stripped Partition' from existing Stripped Partitions}
\usage{
strip_product_r(lhs, rhs)
}
\arguments{
\item{lhs}{Stripped Partition}

\item{rhs}{Stripped Partition}
}
\value{
Stripped Partition
}
\description{
Generate 'Stripped Partition' from existing Stripped Partitions
}
\examples{
\dontrun{
xy <- strip_product_r(x,y)
}
}
