% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-frame_to_pl.R
\name{valid_pl_index}
\alias{valid_pl_index}
\title{Validate if an input object is 'pl_index'}
\usage{
valid_pl_index(obj)
}
\arguments{
\item{obj}{S3 Object}
}
\value{
R Object
}
\description{
Validate if an input object is 'pl_index'
}
\examples{
\dontrun{
test <- valid_pl_index(obj)
}
}
