% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-frame_to_pl.R
\name{valid_pl_map}
\alias{valid_pl_map}
\title{Validate if an input object is 'pl_map'}
\usage{
valid_pl_map(obj)
}
\arguments{
\item{obj}{S3 Object}
}
\value{
R Object
}
\description{
Validate if an input object is 'pl_map'
}
\examples{
\dontrun{
test <- valid_pl_map(s3_obj)
}
}
