% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-attr_set.R
\name{validate_attr_set}
\alias{validate_attr_set}
\title{S3 Validator for Class 'attr_set'}
\usage{
validate_attr_set(obj, bool)
}
\arguments{
\item{obj}{S3 Object}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'attr_set'
}
\examples{
\dontrun{
test <- validate_attr_set(obj, FALSE)
}
}
