% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-fdep.R
\name{validate_fdep}
\alias{validate_fdep}
\title{S3 Validator for Class 'fdep'}
\usage{
validate_fdep(obj, bool)
}
\arguments{
\item{obj}{S3 Object}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'fdep'
}
\examples{
\dontrun{
result <- validate_fdep(obj, TRUE)
}
}
