% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-func_dep.R
\name{validate_func_dep}
\alias{validate_func_dep}
\title{S3 Validator for Class 'func_dep'}
\usage{
validate_func_dep(obj, bool)
}
\arguments{
\item{obj}{S3 Object}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'func_dep'
}
\examples{
\dontrun{
result <- validate_func_dep(obj, TRUE)
}
}
