% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_class-pos_list.R
\name{validate_pos_list}
\alias{validate_pos_list}
\title{S3 Validator for Class 'pos_list'}
\usage{
validate_pos_list(obj, bool)
}
\arguments{
\item{obj}{S3 Object}

\item{bool}{TRUE/FALSE}
}
\value{
S3 Object
}
\description{
S3 Validator for Class 'pos_list'
}
\examples{
\dontrun{
test <- validate_pos_list(s3_obj, FALSE)
}
}
