% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-err_func.R
\name{err_fd_from_records}
\alias{err_fd_from_records}
\title{Calculate 'Functional Dependency Error' from Position List Records}
\usage{
err_fd_from_records(pos, xcol, ycol, err_only)
}
\arguments{
\item{pos}{S3 Object of class 'pos_list'}

\item{xcol}{character - Vector of column names}

\item{ycol}{character - Vector of column names}

\item{err_only}{TRUE/FALSE - Optionally specify if only failing values should be returned}
}
\value{
R Object
}
\description{
Calculate 'Functional Dependency Error' from Position List Records
}
\examples{
\dontrun{
result <- err_key_from_records(pos_list, xcol, ycol)
}
}
