% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-err_func.R
\name{err_key_from_records}
\alias{err_key_from_records}
\title{Calculate 'Key Error' from Position List Records}
\usage{
err_key_from_records(pos, col)

err_key_from_records(pos, col)
}
\arguments{
\item{pos}{S3 Object of class 'pos_list'}

\item{col}{character - Vector of column names}
}
\value{
R Object

R Object
}
\description{
Calculate 'Key Error' from Position List Records

Calculate 'Functional Dependency Error' from Position List Records
}
\examples{
\dontrun{
result <- err_key_from_records(strip_part)
}
\dontrun{
result <- err_key_from_records(strip_part)
}
}
